[Feature #111860601] Developer meetup (Add success notification)
Added a notification to indicate that the meetup creation was
successful. Like the Dribbble one, it redirects the user to the
FAQ page. If it failed, the user is taken back to the meetup creation
page.

At creation, the meetup will automatically be assigned an approved status
of 'false'.

The approved field is a new field for the Meetups table so that migration
will have to be re-run.